requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
